requirements.txt
SQLAlchemy
pymysql
streamlit
python-dotenv
cryptography